# [5.4] Faster implementation for Arr::crossJoin()