--- a/tests/Support/SupportArrTest.php
+++ b/tests/Support/SupportArrTest.php
@@ -38,25 +38,25 @@ class SupportArrTest extends TestCase
     public function testCrossJoin()
     {
         // Single dimension
-        $this->assertEquals(
+        $this->assertSame(
             [[1, 'a'], [1, 'b'], [1, 'c']],
             Arr::crossJoin([1], ['a', 'b', 'c'])
         );
 
         // Square matrix
-        $this->assertEquals(
+        $this->assertSame(
             [[1, 'a'], [1, 'b'], [2, 'a'], [2, 'b']],
             Arr::crossJoin([1, 2], ['a', 'b'])
         );
 
         // Rectangular matrix
-        $this->assertEquals(
+        $this->assertSame(
             [[1, 'a'], [1, 'b'], [1, 'c'], [2, 'a'], [2, 'b'], [2, 'c']],
             Arr::crossJoin([1, 2], ['a', 'b', 'c'])
         );
 
         // 3D matrix
-        $this->assertEquals(
+        $this->assertSame(
             [
                 [1, 'a', 'I'], [1, 'a', 'II'], [1, 'a', 'III'],
                 [1, 'b', 'I'], [1, 'b', 'II'], [1, 'b', 'III'],
@@ -67,17 +67,17 @@ class SupportArrTest extends TestCase
         );
 
         // With 1 empty dimension
-        $this->assertEquals([], Arr::crossJoin([], ['a', 'b'], ['I', 'II', 'III']));
-        $this->assertEquals([], Arr::crossJoin([1, 2], [], ['I', 'II', 'III']));
-        $this->assertEquals([], Arr::crossJoin([1, 2], ['a', 'b'], []));
+        $this->assertSame([], Arr::crossJoin([], ['a', 'b'], ['I', 'II', 'III']));
+        $this->assertSame([], Arr::crossJoin([1, 2], [], ['I', 'II', 'III']));
+        $this->assertSame([], Arr::crossJoin([1, 2], ['a', 'b'], []));
 
         // With empty arrays
-        $this->assertEquals([], Arr::crossJoin([], [], []));
-        $this->assertEquals([], Arr::crossJoin([], []));
-        $this->assertEquals([], Arr::crossJoin([]));
+        $this->assertSame([], Arr::crossJoin([], [], []));
+        $this->assertSame([], Arr::crossJoin([], []));
+        $this->assertSame([], Arr::crossJoin([]));
 
         // Not really a proper usage, still, test for preserving BC
-        $this->assertEquals([[]], Arr::crossJoin());
+        $this->assertSame([[]], Arr::crossJoin());
     }
 
     public function testDivide()