Follow-up to #19236. ping @JosephSilber

Benchmarks, for 1000 iterations:

    Arr::crossJoin([1, 2], ['a', 'b'], ['I', 'II', 'III']);

    current: 429 ms
    new: 30 ms

    Arr::crossJoin([1, 2], ['a', 'b', 'c']);

    current: 205 ms
    new: 18 ms

props Serg for this [implementation](https://stackoverflow.com/questions/6311779/finding-cartesian-product-with-php-associative-arrays/15973172#15973172). (note I removed the `array_filter` line)

Also added a few tests to reinforce BC.