@vlakoff 2017-07-02
Changed tests to `assertSame` for strict array comparisons. This implementation's results have the same internal ordering as previous code.