@thecrypticace 2017-07-03
👍 

"Your reminder that building arrays with reduce, while fun, is accidentally quadratic."
https://twitter.com/AirspeedSwift/status/627469657264553984

(This is about Swift but the same concept applies here as well)