@vlakoff 2017-07-03
I just ran some quick'n'dirty tests, this new implemention seems to be O(n), and it seems the previous implementation was also O(n).

So the new implementation is just constantly faster :)